fix: read runtime projects for provider test
Generated-By: looper 0.10.3 (runner=fixer, agent=codex)
